# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Accounting System",
    "description": "Personal finance tracking with bank statement imports",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/merchants/": {
      "get": {
        "tags": [
          "merchants"
        ],
        "summary": "List Merchants",
        "description": "List all merchants. Optionally include transaction counts.",
        "operationId": "list_merchants_api_merchants__get",
        "parameters": [
          {
            "name": "with_counts",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "With Counts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MerchantResponse"
                  },
                  "title": "Response List Merchants Api Merchants  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "merchants"
        ],
        "summary": "Create Merchant",
        "description": "Create a new merchant.",
        "operationId": "create_merchant_api_merchants__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/merchants/match": {
      "get": {
        "tags": [
          "merchants"
        ],
        "summary": "Match Merchant",
        "description": "Test matching a raw merchant name against known merchants.",
        "operationId": "match_merchant_api_merchants_match_get",
        "parameters": [
          {
            "name": "raw_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Raw Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantMatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/merchants/{merchant_id}": {
      "patch": {
        "tags": [
          "merchants"
        ],
        "summary": "Update Merchant",
        "description": "Update a merchant.",
        "operationId": "update_merchant_api_merchants__merchant_id__patch",
        "parameters": [
          {
            "name": "merchant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Merchant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "merchants"
        ],
        "summary": "Delete Merchant",
        "description": "Delete a merchant.",
        "operationId": "delete_merchant_api_merchants__merchant_id__delete",
        "parameters": [
          {
            "name": "merchant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Merchant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/merchants/{merchant_id}/merge": {
      "post": {
        "tags": [
          "merchants"
        ],
        "summary": "Merge Merchant",
        "description": "Absorb another merchant's patterns into this one, then delete the source.",
        "operationId": "merge_merchant_api_merchants__merchant_id__merge_post",
        "parameters": [
          {
            "name": "merchant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Merchant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantMergeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/merchants/{merchant_id}/transactions": {
      "get": {
        "tags": [
          "merchants"
        ],
        "summary": "Get Merchant Transactions",
        "description": "Get all transactions matching a merchant's name and patterns.",
        "operationId": "get_merchant_transactions_api_merchants__merchant_id__transactions_get",
        "parameters": [
          {
            "name": "merchant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Merchant Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MerchantCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "patterns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Patterns"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "MerchantCreate",
        "description": "Fields for creating a merchant."
      },
      "MerchantMatchResponse": {
        "properties": {
          "canonical_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Name"
          },
          "merchant_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "match_type": {
            "type": "string",
            "title": "Match Type"
          }
        },
        "type": "object",
        "required": [
          "score",
          "match_type"
        ],
        "title": "MerchantMatchResponse",
        "description": "Result of matching a raw merchant name."
      },
      "MerchantMergeRequest": {
        "properties": {
          "source_merchant_id": {
            "type": "integer",
            "title": "Source Merchant Id"
          }
        },
        "type": "object",
        "required": [
          "source_merchant_id"
        ],
        "title": "MerchantMergeRequest",
        "description": "Request to merge another merchant into this one."
      },
      "MerchantResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "patterns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Patterns"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "transaction_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "MerchantResponse",
        "description": "Merchant response."
      },
      "MerchantUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "patterns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patterns"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          }
        },
        "type": "object",
        "title": "MerchantUpdate",
        "description": "Fields for updating a merchant."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}